refactor: separa types em arquivos individuais por domínio
