+++ src/types/artistaModel.ts
@@ -1,0 +1,5 @@
+export interface Artista {
+  id: string;
+  nome: string;
+  generoId: string;
+}
