+++ src/types/canalVendaModel.ts
@@ -1,0 +1,5 @@
+export interface CanalVenda {
+  id: string;
+  nome: string;
+  taxaPadrao: number;
+}
