+++ src/types/compraModel.ts
@@ -1,0 +1,14 @@
+export interface Compra {
+  id: string;
+  clienteId: string;
+  dataCompra: string;
+  fornecedor: string;
+  valorTotal: number;
+}
+
+export interface ItemCompra {
+  id: string;
+  compraId: string;
+  discoId: string;
+  precoCompra: number;
+}
