+++ src/types/generoMusicalModel.ts
@@ -1,0 +1,4 @@
+export interface GeneroMusical {
+  id: string;
+  nome: string;
+}
